+++ api/db.py
@@ -1,6 +1,7 @@
 from sqlalchemy.orm import sessionmaker
+from sqlalchemy.sql import text
 
-from model import Work, Edition, Item, Link, Rights, Identifier
+from model import Work, Edition, Item, Link
 from model.postgres.item import ITEM_LINKS
 from .utils import APIUtils
 
@@ -14,12 +15,11 @@ class DBClient():
 
         session = sessionmaker(bind=self.engine)()
 
-        query = session.query(Work)\
+        return session.query(Work)\
             .join(Edition)\
             .join(Item, ITEM_LINKS, Link)\
-            .filter(Work.uuid.in_(uuids), Edition.id.in_(editionIds))
-
-        return query.all()
+            .filter(Work.uuid.in_(uuids), Edition.id.in_(editionIds))\
+            .all()
 
     def fetchSingleWork(self, uuid):
         session = sessionmaker(bind=self.engine)()
@@ -28,3 +28,22 @@ class DBClient():
             .join(Edition)\
             .join(Item, ITEM_LINKS, Link)\
             .filter(Work.uuid == uuid).first()
+
+    def fetchSingleEdition(self, editionID, showAll=False):
+        session = sessionmaker(bind=self.engine)()
+
+        return session.query(Edition)\
+            .outerjoin(Item, ITEM_LINKS, Link)\
+            .filter(Edition.id == editionID).first()
+
+    def fetchRowCounts(self):
+        session = sessionmaker(bind=self.engine)()
+
+        countQuery = text("""SELECT relname AS table, reltuples AS row_count
+            FROM pg_class c JOIN pg_namespace n ON (n.oid = c.relnamespace)
+            WHERE nspname NOT IN ('pg_catalog', 'information_schema')
+            AND relkind = 'r'
+            AND relname IN ('records', 'works', 'editions', 'items', 'links')
+        """)
+
+        return session.execute(countQuery)